refactor(ricer::config): move production and test code for settings into separate file
